初始化 VSCode 插件项目 #658
